requirements.txt
# Data processing
matplotlib>=3.8.0
rasterstats>=0.19.0
rasterio==1.3.10
geopandas==0.14.4
xarray==2024.2.0
numpy==1.26.4
pandas==2.2.1
netCDF4==1.6.5
h5py==3.10.0
shapely==2.0.3

# HTTP / API
requests==2.31.0
httpx==0.27.0
python-dotenv==1.0.1

# Backend
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Tile server
titiler[full]==0.18.0

# Scheduling
schedule==1.2.1
